Rename math directory -> elementwise
